Minor refactor of @compiled/babel-plugin
