+++ packages/babel-plugin/src/babel-plugin.tsx
@@ -1,16 +1,16 @@
+import { basename } from 'path';
+
 import { declare } from '@babel/helper-plugin-utils';
 import template from '@babel/template';
 import * as t from '@babel/types';
 import jsxSyntax from '@babel/plugin-syntax-jsx';
-import type { NodePath } from '@babel/traverse';
 import { unique } from '@compiled/utils';
-import * as path from 'path';
-import { importSpecifier } from './utils/ast-builders';
-import { Cache } from './utils/cache';
+
+import { visitClassNamesPath } from './class-names';
 import { visitCssPropPath } from './css-prop';
 import { visitStyledPath } from './styled';
-import { visitClassNamesPath } from './class-names';
 import type { State } from './types';
+import { appendRuntimeImports } from './utils/append-runtime-imports';
 import {
   isCompiledCSSCallExpression,
   isCompiledCSSTaggedTemplateExpression,
@@ -19,74 +19,18 @@ import {
   isCompiledStyledCallExpression,
   isCompiledStyledTaggedTemplateExpression,
 } from './utils/ast';
+import { Cache } from './utils/cache';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const pkgJson = require('../package.json');
+const packageJson = require('../package.json');
 
 let globalCache: Cache | undefined;
-
-const parseFilename = (filename: string | undefined) => {
-  if (!filename) {
-    return 'File';
-  }
-
-  return path.basename(filename);
-};
-
-/**
- * Appends runtime import to code. If it is already present, it will append import specifiers
- * to already imported declaration path else it will create fresh import declaration path
- * with runtime import specifiers.
- *
- * @param path ImportDeclaration node path
- */
-const appendRuntimeImports = (path: NodePath<t.ImportDeclaration>) => {
-  const runtimeImportNames = ['ax', 'ix', 'CC', 'CS'];
-  const runtimeImportModuleName = '@compiled/react/runtime';
-
-  // Check if we have any sibling runtime import
-  const runtimeImportFound = path
-    .getAllPrevSiblings()
-    .concat(path.getAllNextSiblings())
-    .find(
-      (path) =>
-        t.isImportDeclaration(path.node) && path.node.source.value === runtimeImportModuleName
-    ) as NodePath<t.ImportDeclaration> | undefined;
-
-  if (runtimeImportFound) {
-    /**
-     * Get local import name instead of imported name to handle scenario when
-     * import specifier is imported as named and normal both.
-     *
-     * eg. import { CC as CompiledRoot, ax, CC, CS } from '@compiled/react/runtime';
-     * In above example `CC` is used both as `CompiledRoot` and `CC`.
-     */
-    const localImportNames = runtimeImportFound
-      .get('specifiers')
-      .map((specifier) => specifier.node.local.name);
-
-    runtimeImportNames.forEach((runtimeImportName) => {
-      // Avoids duplicate imports from being appended if already present
-      if (!localImportNames.includes(runtimeImportName)) {
-        runtimeImportFound.pushContainer('specifiers', importSpecifier(runtimeImportName));
-      }
-    });
-  } else {
-    // Add the runtime entrypoint module
-    path.insertBefore(
-      t.importDeclaration(
-        runtimeImportNames.map((runtimeImportName) => importSpecifier(runtimeImportName)),
-        t.stringLiteral(runtimeImportModuleName)
-      )
-    );
-  }
-};
 
 export default declare<State>((api) => {
   api.assertVersion(7);
 
   return {
-    name: pkgJson.name,
+    name: packageJson.name,
     inherits: jsxSyntax,
     pre() {
       this.sheets = {};
@@ -126,11 +70,12 @@ export default declare<State>((api) => {
             path.unshiftContainer('body', template.ast(`import { forwardRef } from 'react'`));
           }
 
-          const version = process.env.TEST_PKG_VERSION || pkgJson.version;
+          const filename = basename(state.filename ?? '') || 'File';
+          const version = process.env.TEST_PKG_VERSION || packageJson.version;
 
           path.addComment(
             'leading',
-            ` ${parseFilename(state.filename)} generated by ${pkgJson.name} v${version} `
+            ` ${filename} generated by ${packageJson.name} v${version} `
           );
 
           // Add a line break the comment
